requirements.txt
Django>=5.0.0
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
django-celery-beat>=2.5.0
yfinance>=0.2.40
whitenoise>=6.0.0
dj-database-url>=2.0.0
apscheduler>=3.10.0
django-filter>=23.0.0
